# [util] Improve vmem generation performance