--- a/util/design/secded_gen.py
+++ b/util/design/secded_gen.py
@@ -320,15 +320,6 @@ def print_dec(n, k, m, codes, codetype, print_type="logic"):
     return outstr
 
 
-# return whether an integer is a power of 2
-def is_pow2(n):
-    return (n & (n - 1) == 0) and n != 0
-
-
-def is_odd(n):
-    return (n % 2) > 0
-
-
 def verify(cfgs):
     error = 0
     for cfg in cfgs['cfgs']:
@@ -589,46 +580,28 @@ def _inv_hamming_code(k, m):
     return _hamming_code(k, m)
 
 
-# n = total bits
-# k = data bits
-# m = parity bits
-# generate hamming code
+# Generate hamming code.
+# total_cnt = number of total bits
+# data_cnt = number of data bits
+# parity_cnt = number of parity bits
 @functools.lru_cache(maxsize=None)
-def _hamming_code(k, m):
-
-    n = k + m
-
+def _hamming_code(data_cnt, parity_cnt):
     # construct a list of code tuples.
     # Tuple corresponds to each bit position and shows which parity bit it participates in
     # Only the data bits are shown, the parity bits are not.
-    codes = []
-    for pos in range(1, n + 1):
-        # this is a valid parity bit position or the final parity bit
-        if (is_pow2(pos) or pos == n):
-            continue
-        else:
-            code = ()
-            for p in range(m):
-
-                # this is the starting parity position
-                parity_pos = 2**p
-
-                # back-track to the closest parity bit multiple and see if it is even or odd
-                # If even, we are in the skip phase, do not include
-                # If odd, we are in the include phase
-                parity_chk = int((pos - (pos % parity_pos)) / parity_pos)
-
-                # valid for inclusion or final parity bit that includes everything
-                if is_odd(parity_chk) or p == m - 1:
-                    code = code + (p, )
-
-            codes.append(code)
-
-    # final parity bit includes all ECC bits
-    for p in range(m - 1):
-        codes.append((m - 1, ))
-
-    log.info("Hamming codes {}".format(codes))
+    total_cnt = data_cnt + parity_cnt
+    last_parity = parity_cnt - 1
+    # Data bits.
+    # This excludes bit 0, the last bit, and any bit whose index is a power of 2.
+    data_bits = [b for b in range(1, total_cnt) if b & (b - 1)]
+    # Include the closest previous parity bit if it's odd.
+    # Also include the final parity bit since it includes everything.
+    codes = [
+        tuple(p for p in range(last_parity) if (b >> p) & 1) + (last_parity,) for b in data_bits
+    ]
+    # Final parity bit includes all ECC bits.
+    codes += [(last_parity,)] * last_parity
+    log.info("Hamming codes %s", codes)
     return codes