This PR speeds up vmem generation (mostly ECC encoding) by:
* Moving config HJSON file parsing outside so that it's not loaded repeatedly in tight loops,
* Adding memoization to two functions on the hot path, and
* Rewriting `_hamming_code()` with integer operations and list comprehensions (the original code used floating pointer operations and a more complex nested loop structure).

This speeds up the vmem generation part of `bazel build //sw/device/...`  by ~3x on my machine (~4 mins down from ~11 mins when compared to `master`, which includes the changes in #15387).

```
# upstream/master @ 35b088540
INFO: Elapsed time: 667.718s, Critical Path: 50.85s

# alphan/vmem-speed
INFO: Elapsed time: 236.404s, Critical Path: 25.25s
```

Signed-off-by: Alphan Ulusoy <alphan@google.com>