@drewmacrae 2022-10-12
How did you isolate the vmem portion for timing?