@drewmacrae 2022-10-13
This may be causing an issue with the nightly regression tests. I'm seeing this in the logs `[768 / 981] Action hw/ip/otp_ctrl/data/img_rma.24.vmem; 1403s processwrapper-sandbox ... (44 actions running)`